renamed storage services to configs
